Refactor: create PlayerArea WallArea DiscardArea
