+++ src/css/app.scss
@@ -35,3 +35,25 @@
 .area-field {
   background-color: $green-2;
 }
+
+@for $i from 1 through 24 {
+  .w-#{$i} {
+    width: calc($i / 24) * 100%;
+    height: 100%;
+  }
+}
+
+@for $i from 1 through 24 {
+  .h-#{$i} {
+    height: calc($i / 24) * 100%;
+    width: 100%;
+  }
+}
+
+.w-full {
+  width: 100%;
+}
+
+.h-full {
+  height: 100%;
+}
