+++ src/example/components/DiscardArea.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="column flex-center bg-indigo-4">
+    <div>Discard Area</div>
+  </div>
+</template>
+
+<script lang="ts">
+export default { name: "DiscardArea" };
+</script>
+
+<script setup lang="ts">
+//
+</script>
+
+<style scoped></style>
